[FL-1940] SubGuz: fix te pricenton protocol #755

+++ lib/subghz/protocols/subghz_protocol_princeton.h
@@ -4,7 +4,7 @@
 
 struct SubGhzDecoderPrinceton {
     SubGhzProtocolCommon common;
-    uint16_t te;
+    uint32_t te;
 };
 
 /** SubGhzEncoderPrinceton anonymous type */
